requirements.txt
Flask
flask-cors
tensorflow==2.18.0
numpy
opencv-python==4.10.0.84
pillow